+++ Tests/NanoRoute.Tests/RouterBuilderTests.Endpoints.cs
@@ -40,7 +40,7 @@ namespace NanoRoute.Tests
 
             Assert.That(response.StatusCode, Is.EqualTo(HttpStatusCode.OK));
             Assert.That(await response.Content.ReadAsStringAsync(), Is.EqualTo("exact"));
-            Assert.That(wrongPath.Data[NanoRouteExceptionExtensions.StatusName], Is.EqualTo(HttpStatusCode.NotFound));
+            Assert.That(wrongPath.Status, Is.EqualTo(HttpStatusCode.NotFound));
         }
 
         [Test]
@@ -68,7 +68,7 @@ namespace NanoRoute.Tests
 
             Assert.That(response.StatusCode, Is.EqualTo(HttpStatusCode.OK));
             Assert.That(await response.Content.ReadAsStringAsync(), Is.EqualTo("/files/path/to/file"));
-            Assert.That(wrongPath.Data[NanoRouteExceptionExtensions.StatusName], Is.EqualTo(HttpStatusCode.NotFound));
+            Assert.That(wrongPath.Status, Is.EqualTo(HttpStatusCode.NotFound));
         }
 
         [Test]
@@ -102,7 +102,7 @@ namespace NanoRoute.Tests
 
             Assert.That(await getResponse.Content.ReadAsStringAsync(), Is.EqualTo("GET"));
             Assert.That(await postResponse.Content.ReadAsStringAsync(), Is.EqualTo("POST"));
-            Assert.That(deleteResponse.Data[NanoRouteExceptionExtensions.StatusName], Is.EqualTo(HttpStatusCode.NotFound));
+            Assert.That(deleteResponse.Status, Is.EqualTo(HttpStatusCode.NotFound));
         }
 
         [Test]